requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-config-url==0.1.1
Django==2.0.4
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-node==4.0.0
django-webpack==3.1.1
django-webpack-loader==0.6.0
djangorestframework==3.8.0
idna==2.6
Markdown==2.6.11
oauthlib==2.0.7
optional-django==0.1.0
psycopg2-binary==2.7.4
PyJWT==1.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
urllib3==1.22
